i12e: 'net.go' merged into 'mesh.go' and deleted
